+++ src/App.jsx
@@ -1,30 +1,32 @@
-import React from 'react';
+import React, { useEffect } from 'react';
 import { HashRouter as Router, Routes, Route, } from "react-router-dom";
+import ReactGA from 'react-ga4';
 import Home from './pages/Home';
 import About from './pages/About';
 import Twitch from './pages/Twitch';
-import Terms from './pages/Terms';
-import PrivacyPolicy from './pages/PrivacyPolicy';
 import FourOhFour from './pages/FourOhFour';
 import ComingSoon from './components/ComingSoon/ComingSoon';
+import Analytics from './components/Analytics/Analytics';
 
 const App = () => {
-  const isMaintenance = true;
+  const isMaintenance = false;
+
+  useEffect(() => {
+    ReactGA.initialize(import.meta.env.VITE_ANALYTICS_ID);
+  },[]);
 
   return (
     <>
 
-    { isMaintenance ? <ComingSoon /> : (
-      
+    { isMaintenance ? <ComingSoon /> : (     
       
       <Router>
+        <Analytics />
         <Routes>
           <Route index element={<Home />} />
           <Route exact path="/" element={<Home />} />
           <Route path="/about" element={<About />} />
           <Route path="/twitch" element={<Twitch />} />
-          <Route path='/terms-of-service' element={<Terms />} />
-          <Route path='/privacy-policy' element={<PrivacyPolicy />} />
           <Route path="*" element={<FourOhFour />} />
         </Routes>
       </Router>