+++ src/components/Hero/Hero.jsx
@@ -9,7 +9,7 @@ const Hero = () => {
     <section className="hero" id="hero">
         <div className="decor">
             <div className="code-wrapper">
-                <div className="pseudocode">&#x3C;select id=&#x22;intl-1033&#x22; name=&#x22;country_code&#x22;&#x3E;&#x3C;option value=&#x22;EN&#x22;&#x3E;English&#x3C;/option&#x3E;&#x3C;/select&#x3E;</div>
+                <div className="pseudocode">&#x3C;header class=&#x22;global&#x22;&#x3E;&#x3C;nav&#x3E;&#x3C;ul class=&#x22;menu-toplevel&#x22;&#x3E;&#x3C;li&#x3E;&#x3C;a href=&#x22;#&#x22;&#x3E;Home&#x3C;/a&#x3E;&#x3C;/li&#x3E;&#x3C;/ul&#x3E;&#x3C;/header&#x3E;</div>
             </div>
             <div className="code-wrapper">
                 <div className="pseudocode reverse">&#x66;&#x6F;&#x72;&#x28;&#x24;&#x69;&#x3D;&#x30;&#x3B;&#x24;&#x69;&#x3C;&#x63;&#x6F;&#x75;&#x6E;&#x74;&#x28;&#x24;&#x7A;&#x6C;&#x69;&#x73;&#x74;&#x29;&#x3B;&#x24;&#x69;&#x2B;&#x2B;&#x29;&#x7B;&#x20;&#x65;&#x63;&#x68;&#x6F;&#x20;&#x22;&#x70;&#x61;&#x72;&#x61;&#x6D;&#x65;&#x74;&#x65;&#x72;&#x20;&#x22;&#x2E;&#x28;&#x24;&#x69;&#x2B;&#x31;&#x29;&#x2E;&#x20;&#x24;&#x7A;&#x6C;&#x69;&#x73;&#x74;&#x5B;&#x24;&#x69;&#x5D;&#x3B;&#x20;&#x7D;</div>
@@ -28,7 +28,7 @@ const Hero = () => {
                 also known as <span className="highlight">@the13thgeek</span>!</p>
                 <div className="call-to-action">
                     <NavLink to="/about" className="cta-link">About Me <i className="fa-solid fa-chevron-right"></i></NavLink>
-                    <AnchorLink offset={50} className="cta-link" href='#contact'>Connect <i className="fa-solid fa-chevron-right"></i></AnchorLink>
+                    <AnchorLink offset={50} className="cta-link" href='#connect'>Connect <i className="fa-solid fa-chevron-right"></i></AnchorLink>
                 </div>
             </div>